requirements.txt
agno==1.5.10
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asyncer==0.0.8
attrs==25.3.0
backoff==2.2.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorlog==6.9.0
coolname==2.2.0
datasets==3.6.0
decorator==5.2.1
deepeval==3.0.8
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docstring-parser==0.16
dspy==2.6.27
dspy-ai==2.6.27
execnet==2.1.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
gitpython==3.1.44
google-auth==2.40.3
google-genai==1.19.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.10
importlib-metadata==8.7.0
iniconfig==2.1.0
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json-repair==0.46.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langfuse==3.0.1
litellm==1.72.2
magicattr==0.1.6
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
monotonic==1.6
multidict==6.4.4
multiprocess==0.70.16
nanoid==2.0.0
nest-asyncio==1.6.0
numpy==2.3.0
ollama==0.5.1
openai==1.85.0
openinference-instrumentation==0.1.32
openinference-instrumentation-haystack==0.1.24
openinference-instrumentation-langchain==0.1.43
openinference-instrumentation-openai==0.1.30
openinference-semantic-conventions==0.1.18
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
optuna==4.3.0
packaging==24.2
pandas==2.3.0
pluggy==1.6.0
portalocker==3.1.1
posthog==3.25.0
propcache==0.3.2
protobuf==5.29.5
py==1.11.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
pyfiglet==1.0.3
pygments==2.19.1
pytest==8.4.0
pytest-asyncio==1.0.0
pytest-repeat==0.9.4
pytest-rerunfailures==12.0
pytest-xdist==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
retry==0.9.2
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
sentry-sdk==2.29.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.41
tabulate==0.9.0
tenacity==9.0.0
termcolor==3.1.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
